Change embedding used to Wikipedia + Gigaword
While the Google News embedding trained on 3M words with 300 dimensions
might end up giving the best results, it's gigantic. At least for the
time being while iterating and trying to get a model up and running,
it'll better to use the smaller Wikipedia + Gigaword embedding. It also
happens to be directly downloadable via a URL, which is why it's been
added to the Makefile (the Google News one has to pass through your
Google Drive).