Build basic embedding model; adjust preprocessing
The building of a basic LSTM with a simple embedding layer led to
playing around with the preprocessing step and determining that it
didn't save time to perform the preprocessing ahead of time and save
the results. Instead, it was faster to perform it in the `__main__`
block of the `lstm.py`. This led to the changes of functionalizing the
reading in of the data in `preprocessing.py` and the moving of the
preprocessing to the `__main__` block of `lstm.py`.