Build function to filter corpus
For iterative processes, it'd be nice to filter the embedding corpus to only
those words that we observe in our articles. Theoretically this would
speed up iterations, which will be helpful when initially building up the
LSTM and getting the artchitecture all set. In practice, thus far the
speed gains are not there.

After some minor tests, (5 runs w and w/o this change), the generation
of all of the lookup dictionaries and the small predict step are slower.
However, I imagine that the large speed gains will come in the fit step.
Further tests will be done when that comes up, and if there are still no
speed gains this code will be removed in a later refactor.